# ordentshop-apidoc.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "ORDENTSHOP-API",
      "description": "API Documentation",
      "contact": {
        "name": "Ordentshop"
      },
      "license": {
        "name": "Agus Priyatin",
        "url": "https://aguspriyatin.wordpress.com/kontak"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "https://ordentshop-api.herokuapp.com/api/v1",
        "variables": {}
      }
    ],
    "paths": {
...
      "/users/{id}": {
        "get": {
          "tags": [
            "users"
          ],
          "summary": "users",
          "operationId": "Getusers",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        }
      },
      "/users/login": {
        "post": {
          "tags": [
            "users"
          ],
          "summary": "login",
          "operationId": "login",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginrequest"
                },
                "example": {
                  "email": "agus3@gmail.com",
                  "password": "Agus1234"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        }
      },
...
      "/users/{userId}": {
        "delete": {
          "tags": [
            "users"
          ],
          "summary": "users",
          "operationId": "Deleteusers",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "loginrequest": {
          "title": "loginrequest",
          "required": [
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "example": {
            "email": "agus3@gmail.com",
            "password": "Agus1234"
          }
        }
      },
      "securitySchemes": {
        "bearer": {
          "type": "https",
          "scheme": "bearer"
        }
      }
    },
    "security": [],
...
  }